requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.1
Flask==3.1.1
Flask-SocketIO==5.5.1
greenlet==3.2.3
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyserial==3.5
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0